finance_backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
httpx
pytest
pytest-asyncio
